+++ Sources/P5Swift/Core/P5SketchInternal.swift
@@ -1,8 +1,0 @@
-@MainActor
-protocol P5SketchInternal {
-    var isLooping: Bool { get set }
-    var framesPerSecond: Double { get set }
-    var userWantsRedraw: Bool { get set }
-
-    func addOperation(_ operation: P5Operation)
-}